refact: organiza o script de rebuild

+++ src/rebuild.sh
@@ -1,8 +1,27 @@
 #!/bin/bash
 
-rm app.db
-rm -rf migrations
+if [ -e app.db ]; then
+    rm app.db
+fi
 
-flask db init
-flask db migrate
-flask db upgrade
+if [ -d migrations2 ]; then
+    rm -rf migrations
+fi
+
+migrate_models() {
+    printf "\n\n\nMigrating models...\n\n\n"
+    flask db init
+}
+
+generate_revision() {
+    printf "\n\n\nGenerating new revision...\n\n\n"
+    flask db migrate
+}
+
+upgrade_version() {
+    printf "\n\n\nUpgrading to the new version...\n\n\n"
+    flask db upgrade
+}
+
+migrate_models && generate_revision && upgrade_version
+printf "\n\n\n"
